(get|put|copy_from|copy_to)_user turn out to be very slow done one at a time, especially on AMD2950x.  Solution is the same as overarching idea of this module, namely to do more real work all at once.
So, we kmalloc (being careful to zero and not leak any kernel memory back
to user space) a big buffer, copy the calls[] & put the rets[] there, and
at the end copy all needed back to user space.  The conception here of
the (slightly) self-modifying program (maybe self-parameterized program?)
here means the "real" program in user space is unaltered which is maybe
a good feature (no select(2)-like problem requiring fdset restoration).
As part of this we rename `a0` to the more descriptive `ur0` & introduce
`kc0`, `urEnd` and so on for bounds checking and translating addresses
the user specifies to the new in-kernel memory buffer.

This really speeds things up.  On a 3.8 GHz AMD 2950x, times go from
108 ns/batch element (where BATCH_EMUL=1 ./unimpl ran 1.7x *faster* than
sys_batch-using ./unimpl!) down to 24 ns or almost 5x faster.  This new
version now makes unimpl take about ns = 242 + 26*n giving batching 9x
boost potential for the cheapest syscalls on the CPU I have with the
fastest syscall transitions and slowest user<->kernel copies.  On an
ancient i5-540M, times also improve from 128 ns/elt to 35 ns/elt or by
3.7x.  The least seen improvement was over on a i7-6700k@4.7GHz (with
kernel PTI active) where ns goes from 358 + 48.6*n down to 467.6 +
23.8*n.  That needs about (467.6-358)/(48.6-23.8) or 4.4 elements per
batch to justify the kmalloc/kfree overheads.  Common case batch sizes
should be much more than 4.4 elements if there was any motivation for
batching anyway.

So, it makes sense to not condition this optimization on the batch size
unless there is some CPU with hyper fast user<->kernel data motion.
(It's also nice to leave the user buffer alone for use in a loop, as
mentioned in the second paragraph.)